refactor: reduce large files
