+++ ui/src/pages/targeting/index.tsx
@@ -8,40 +8,22 @@ import {
   useImperativeHandle,
   useRef,
 } from 'react';
-import {
-  Form,
-  Radio,
-  CheckboxProps,
-  TextAreaProps,
-  InputOnChangeData,
-  Loader,
-  Dropdown,
-  DropdownItemProps,
-  Popup,
-  RadioProps,
-} from 'semantic-ui-react';
-import { useParams, useHistory, Prompt } from 'react-router-dom';
+import { Form, Radio, CheckboxProps, Loader } from 'semantic-ui-react';
+import { Prompt } from 'react-router-dom';
 import isEqual from 'lodash/isEqual';
 import cloneDeep from 'lodash/cloneDeep';
 import moment from 'moment';
-import { useForm } from 'react-hook-form';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { v4 as uuidv4 } from 'uuid';
-import { diff } from 'deep-diff';
-import Joyride, { CallBackProps, EVENTS, ACTIONS, Step } from 'react-joyride';
 import Rules from './components/Rules';
 import DefaultRule from './components/DefaultRule';
 import DisabledServe from './components/DisabledServe';
 import { useBeforeUnload } from './hooks';
-import Icon from 'components/Icon';
 import message from 'components/MessageBox';
-import Modal from 'components/Modal';
 import Button from 'components/Button';
 import Variations from 'components/Variations';
 import SectionTitle from 'components/SectionTitle';
 import EventTracker from 'components/EventTracker';
-import Diff from 'components/Diff';
-import { approveToggle, saveToggle } from 'services/toggle';
 import { replaceSpace } from 'utils/tools';
 import {
   variationContainer,
@@ -55,29 +37,20 @@ import { VariationColors } from 'constants/colors';
 import {
   IApprovalInfo,
   ICondition,
-  IDictionary,
-  IOption,
   IRule,
-  IServe,
   ITarget,
   ITargeting,
-  ITargetingParams,
   IToggleInfo,
   IVariation,
 } from 'interfaces/targeting';
-import { IRouterParams } from 'interfaces/project';
 import { ISegmentList } from 'interfaces/segment';
 import { DATETIME_TYPE, SEGMENT_TYPE } from 'components/Rule/constants';
-import { DiffServe } from 'components/Diff/DiffServe';
-import VariationsDiffContent from 'components/Diff/VariationsDiffContent';
-import { I18NRules, RulesDiffContent } from 'components/Diff/RulesDiffContent';
-import { DiffStatusContent } from 'components/Diff/DiffStatus';
 import SizeTips from 'components/SizeTips';
-import { commonConfig, floaterStyle, tourStyle } from 'constants/tourConfig';
-import { getFromDictionary, saveDictionary } from 'services/dictionary';
-import { USER_GUIDE_LAYOUT, USER_GUIDE_TARGETING } from 'constants/dictionary_keys';
+import UserGuide from './components/UserGuide';
+import PublishModal from './components/PublishModal';
 import { useFormErrorScrollIntoView } from 'hooks';
 import { getBytes, MAX_SIZE } from 'utils/bytes';
+
 import styles from './index.module.scss';
 
 interface IProps {
@@ -94,50 +67,6 @@ interface IProps {
   initTargeting(): void;
   saveToggleDisable(status: boolean): void;
 }
-
-const STEPS: Step[] = [
-  {
-    content: (
-      <div>
-        <div className="joyride-title">
-          <FormattedMessage id="guide.toggle.targeting.step1.title" />
-        </div>
-        <ul className="joyride-item">
-          <li>
-            <FormattedMessage id="guide.toggle.targeting.step1.off" />
-          </li>
-          <li>
-            <FormattedMessage id="guide.toggle.targeting.step1.on" />
-          </li>
-        </ul>
-        <div className="joyride-pagination">1/2</div>
-      </div>
-    ),
-    spotlightPadding: 20,
-    placement: 'right',
-    target: '.joyride-toggle-status',
-    ...commonConfig,
-  },
-  {
-    content: (
-      <div>
-        <div className="joyride-title">
-          <FormattedMessage id="guide.toggle.targeting.step2.title" />
-        </div>
-        <ul className="joyride-item">
-          <li>
-            <FormattedMessage id="guide.toggle.targeting.step2.default" />
-          </li>
-        </ul>
-        <div className="joyride-pagination">2/2</div>
-      </div>
-    ),
-    spotlightPadding: 4,
-    placement: 'right',
-    target: '.joyride-default-rule',
-    ...commonConfig,
-  },
-];
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 const Targeting = forwardRef((props: IProps, ref: any) => {
@@ -159,33 +88,17 @@ const Targeting = forwardRef((props: IProps, ref: any) => {
   const { variations, saveVariations } = variationContainer.useContainer();
   const { defaultServe, saveDefaultServe } = defaultServeContainer.useContainer();
   const { disabledServe, saveDisabledServe } = disabledServeContainer.useContainer();
-  const { projectKey, environmentKey, toggleKey } = useParams<IRouterParams>();
   const [open, setOpen] = useState<boolean>(false);
   const [publishDisabled, setPublishDisabled] = useState<boolean>(true);
   const [publishTargeting, setPublishTargeting] = useState<ITargeting>();
-  const [comment, saveComment] = useState<string>('');
   const [isLoading, setLoading] = useState<boolean>(false);
-  const [options, saveOptions] = useState<IOption[]>();
-  const [run, saveRun] = useState<boolean>(false);
-  const [stepIndex, saveStepIndex] = useState<number>(0);
-  const [isCollect, saveIsCollect] = useState<string>('');
-  const [isDiffChange, saveIsDiffChange] = useState<boolean>(false);
   const [size, saveSize] = useState<number>(0);
   const [sizeConfirm, setSizeConfirm] = useState<boolean>(false);
-  const history = useHistory();
   const intl = useIntl();
   const formRef = useRef();
 
   useBeforeUnload(!publishDisabled, intl.formatMessage({ id: 'targeting.page.leave.text' }));
   useImperativeHandle(ref, () => publishDisabled, [publishDisabled]);
-
-  const {
-    formState: newFormState,
-    trigger: newTrigger,
-    register: newRegister,
-    setValue: newSetValue,
-    clearErrors,
-  } = useForm();
 
   const {
     setValue,
@@ -216,31 +129,6 @@ const Targeting = forwardRef((props: IProps, ref: any) => {
       });
     });
   }, [setBeforeScrollCallback, rules, saveRules]);
-
-  useEffect(() => {
-    Promise.all([
-      getFromDictionary<IDictionary>(USER_GUIDE_LAYOUT),
-      getFromDictionary<IDictionary>(USER_GUIDE_TARGETING),
-    ]).then((res) => {
-      // After finishing layout user guide, then show targeting user guide
-      if (res[0].success && res[0].data && parseInt(JSON.parse(res[0].data.value)) === 3) {
-        const { success, data } = res[1];
-        if (success && data) {
-          const savedData = JSON.parse(data.value);
-          if (parseInt(savedData) !== STEPS.length) {
-            setTimeout(() => {
-              saveRun(true);
-            }, 0);
-            saveStepIndex(parseInt(savedData));
-          }
-        } else {
-          setTimeout(() => {
-            saveRun(true);
-          }, 0);
-        }
-      }
-    });
-  }, []);
 
   useEffect(() => {
     if (targeting) {
@@ -371,27 +259,6 @@ const Targeting = forwardRef((props: IProps, ref: any) => {
     }
   }, [publishTargeting, initialTargeting]);
 
-  useEffect(() => {
-    const options = approvalInfo?.reviewers?.map((name: string) => {
-      return {
-        key: name,
-        value: name,
-        text: name,
-      };
-    });
-
-    saveOptions(options);
-  }, [approvalInfo]);
-
-  useEffect(() => {
-    newRegister('reason', {
-      required: approvalInfo?.enableApproval,
-    });
-    newRegister('radioGroup', {
-      required: !approvalInfo?.enableApproval && allowEnableTrackEvents,
-    });
-  }, [newRegister, approvalInfo, allowEnableTrackEvents]);
-
   const validateForm = useCallback(() => {
     let isError = false;
 
@@ -425,23 +292,23 @@ const Targeting = forwardRef((props: IProps, ref: any) => {
     return isError;
   }, [intl, rules, setError, toggleInfo, variations]);
 
-    //Check if the size of the data is too large
-    useEffect(() => {
-      if(publishTargeting) {
-        const bytes = getBytes(JSON.stringify(publishTargeting.content));
-        saveSize(bytes);
-      }
-    }, [publishTargeting]);
-  
-    const sizeState = useMemo(() => {
-      let state = 'normal';
-      if(size >= MAX_SIZE) {
-        state = 'large';
-      } else if(size > MAX_SIZE * 0.8) {
-        state = 'near';
-      }
-      return state;
-    }, [size]);
+  //Check if the size of the data is too large
+  useEffect(() => {
+    if(publishTargeting) {
+      const bytes = getBytes(JSON.stringify(publishTargeting.content));
+      saveSize(bytes);
+    }
+  }, [publishTargeting]);
+
+  const sizeState = useMemo(() => {
+    let state = 'normal';
+    if(size >= MAX_SIZE) {
+      state = 'large';
+    } else if(size > MAX_SIZE * 0.8) {
+      state = 'near';
+    }
+    return state;
+  }, [size]);
 
   const onSubmit = useCallback(() => {
     if (validateForm()) {
@@ -459,205 +326,11 @@ const Targeting = forwardRef((props: IProps, ref: any) => {
     scrollToError();
   }, [scrollToError, validateForm]);
 
-  const handlePublishCancel = useCallback(() => {
-    setOpen(false);
-    saveComment('');
-    saveIsCollect('');
-    newSetValue('reason', '');
-    newSetValue('radioGroup', '');
-    clearErrors();
-  }, [newSetValue, clearErrors]);
-
-  const handlePublishConfirm = useCallback(async () => {
-    if (approvalInfo?.enableApproval && comment === '') {
-      await newTrigger('reason');
-      return;
-    }
-
-    if (!approvalInfo?.enableApproval && isCollect === '' && allowEnableTrackEvents) {
-      await newTrigger('radioGroup');
-      return;
-    }
-    setOpen(false);
-    setLoading(true);
-    if (publishTargeting) {
-      let res;
-      if (approvalInfo && approvalInfo.enableApproval) {
-        res = await approveToggle(projectKey, environmentKey, toggleKey, {
-          ...publishTargeting,
-          comment,
-          reviewers: approvalInfo.reviewers,
-          baseVersion: latestVersion,
-        });
-      } else {
-        const params: ITargetingParams = {
-          ...publishTargeting,
-          comment,
-          baseVersion: latestVersion,
-        };
-
-        if (isCollect === 'yes') {
-          params.trackAccessEvents = true;
-        }
-        res = await saveToggle(projectKey, environmentKey, toggleKey, params);
-      }
-      if (res.success) {
-        if (approvalInfo && approvalInfo?.enableApproval) {
-          message.success(intl.formatMessage({ id: 'targeting.approval.request.success' }));
-        } else {
-          message.success(intl.formatMessage({ id: 'targeting.publish.success.text' }));
-        }
-        initTargeting();
-        saveComment('');
-        saveIsCollect('');
-      } else {
-        message.error(res.message || intl.formatMessage({ id: 'targeting.publish.error.text' }));
-      }
-      newSetValue('reason', '');
-      newSetValue('radioGroup', '');
-      setLoading(false);
-    }
-  }, [
-    intl,
-    approvalInfo, 
-    comment, 
-    isCollect, 
-    allowEnableTrackEvents, 
-    publishTargeting, 
-    projectKey, 
-    environmentKey, 
-    toggleKey, 
-    latestVersion,
-    newTrigger, 
-    newSetValue, 
-    initTargeting, 
-  ]);
-
   const disabledText = useMemo(() => {
     if (variations[disabledServe.select]) {
       return variations[disabledServe.select].name || variations[disabledServe.select].value;
     }
   }, [disabledServe.select, variations]);
-
-  const handleInputComment = useCallback((e: SyntheticEvent, data: TextAreaProps | InputOnChangeData) => {
-    saveComment(data.value as string);
-  }, []);
-
-  const renderLabel = useCallback((label: DropdownItemProps) => {
-    return {
-      content: label.text,
-      removeIcon: null,
-    };
-  }, []);
-
-  const handleGotoSetting = useCallback(() => {
-    history.push(`/${projectKey}/${environmentKey}/settings`);
-  }, [history, projectKey, environmentKey]);
-
-  const handleJoyrideCallback = useCallback((data: CallBackProps) => {
-    const { action, index, type } = data;
-
-    if (([EVENTS.STEP_AFTER, EVENTS.TARGET_NOT_FOUND] as string[]).includes(type)) {
-      const nextStepIndex = index + (action === ACTIONS.PREV ? -1 : 1);
-      saveStepIndex(nextStepIndex);
-      saveDictionary(USER_GUIDE_TARGETING, nextStepIndex);
-    }
-  }, []);
-
-  const preDiffServe = useCallback((serve?: IServe, variations?: IVariation[]) => {
-    if (!serve || !variations) {
-      return;
-    }
-    const obj: {
-      select?: string;
-      split?: string[];
-    } = {};
-    if (serve.select !== undefined && typeof serve.select === 'number') {
-      obj.select = variations[serve.select].name;
-    }
-    if (serve.split !== undefined) {
-      obj.split = serve.split.map((item: number, index: number) => {
-        return `${variations[index].name}: ${item / 100}%`;
-      });
-    }
-    return obj;
-  }, []);
-
-  const beforeServeDiff = useCallback(
-    (before, after) => {
-      const left = before;
-      const right = after;
-
-      return [
-        preDiffServe(left, initialTargeting?.content.variations),
-        preDiffServe(right, publishTargeting?.content.variations),
-      ];
-    },
-    [preDiffServe, initialTargeting?.content.variations, publishTargeting?.content.variations]
-  );
-
-  const beforeRuleDiff = useCallback(
-    (before, after) => {
-      const left = I18NRules(before, intl);
-      const right = I18NRules(after, intl);
-      return [
-        left.map((item: IRule) => {
-          return {
-            ...item,
-            serve: preDiffServe(item.serve, initialTargeting?.content.variations),
-          };
-        }),
-        right.map((item: IRule) => {
-          return {
-            ...item,
-            serve: preDiffServe(item.serve, publishTargeting?.content.variations),
-          };
-        }),
-      ];
-    },
-    [preDiffServe, initialTargeting?.content.variations, publishTargeting?.content.variations, intl]
-  );
-
-  const handleRadioChange = useCallback((e: SyntheticEvent, detail: RadioProps) => {
-    saveIsCollect(detail.value as string);
-  }, []);
-
-  useEffect(() => {
-    if (initialTargeting && publishTargeting && open) {
-      let isDiffChange = false;
-      if (
-        diff(initialTargeting.disabled, publishTargeting.disabled) ||
-        diff(initialTargeting.content.defaultServe, publishTargeting.content.defaultServe) ||
-        diff(initialTargeting.content.disabledServe, publishTargeting.content.disabledServe)
-      ) {
-        isDiffChange = true;
-      }
-
-      const diffVariation = diff(initialTargeting.content.variations, publishTargeting.content.variations);
-      if (diffVariation) {
-        diffVariation.forEach(item => {
-          if (item.kind !== 'E') {
-            isDiffChange = true;
-          } else if (item.kind === 'E' && (item.path?.[1] !== 'name' && item.path?.[1] !== 'description')) {
-            isDiffChange = true;
-          }
-        });
-      }
-
-      const diffRules = diff(initialTargeting.content.rules, publishTargeting.content.rules);
-      if (diffRules) {
-        diffRules.forEach(item => {
-          if (item.kind !== 'E') {
-            isDiffChange = true;
-          } else if (item.kind === 'E' && item.path?.[1] !== 'name') {
-            isDiffChange = true;
-          }
-        });
-      }
-
-      saveIsDiffChange(isDiffChange);
-    }
-  }, [open, beforeRuleDiff, initialTargeting, publishTargeting]);
 
   return (
     <>
@@ -742,208 +415,20 @@ const Targeting = forwardRef((props: IProps, ref: any) => {
             </Button>
           </EventTracker>
         </div>
-        <Modal open={open} width={800} handleCancel={handlePublishCancel} handleConfirm={handlePublishConfirm}>
-          <div>
-            <div className={styles['modal-header']}>
-              <span className={styles['modal-header-text']}>
-                <FormattedMessage id="targeting.publish.modal.title" />
-              </span>
-              <Icon customclass={styles['modal-close-icon']} type="close" onClick={handlePublishCancel} />
-            </div>
-            <div className={styles['modal-content']}>
-              {
-                isDiffChange && trackEvents && (
-                  <div className={styles['publish-tips']}>
-                    <Icon type='error-circle' customclass={styles['error-circle']} />
-                    <FormattedMessage id='targeting.publish.tips' />
-                  </div>
-                )
-              }
-              <Diff
-                sections={[
-                  {
-                    before: {
-                      disabled: initialTargeting?.disabled,
-                    },
-                    after: {
-                      disabled: publishTargeting?.disabled,
-                    },
-                    title: intl.formatMessage({ id: 'targeting.status.text' }),
-                    renderContent: (content) => {
-                      return <DiffStatusContent content={content} />;
-                    },
-                    diffKey: 'status',
-                  },
-                  {
-                    before: initialTargeting?.content.variations,
-                    after: publishTargeting?.content.variations,
-                    title: intl.formatMessage({ id: 'common.variations.text' }),
-                    renderContent: (content) => {
-                      return <VariationsDiffContent content={content} />;
-                    },
-                    diffKey: 'variations',
-                  },
-                  {
-                    before: initialTargeting?.content.rules,
-                    after: publishTargeting?.content.rules,
-                    title: intl.formatMessage({ id: 'common.rules.text' }),
-                    renderContent: (content) => {
-                      return <RulesDiffContent content={content} />;
-                    },
-                    beforeDiff: beforeRuleDiff,
-                    diffKey: 'rules',
-                  },
-                  {
-                    before: initialTargeting?.content.defaultServe,
-                    after: publishTargeting?.content.defaultServe,
-                    title: intl.formatMessage({ id: 'targeting.default.rule' }),
-                    renderContent: (content) => {
-                      return <DiffServe content={content} />;
-                    },
-                    beforeDiff: beforeServeDiff,
-                    diffKey: 'default',
-                  },
-                  {
-                    title: intl.formatMessage({ id: 'common.disabled.return.type.text' }),
-                    before: initialTargeting?.content.disabledServe,
-                    after: publishTargeting?.content.disabledServe,
-                    renderContent: (content) => {
-                      return <DiffServe content={content} />;
-                    },
-                    beforeDiff: beforeServeDiff,
-                    diffKey: 'disabled',
-                  },
-                ]}
-                maxHeight={341}
-              />
-              <div className={styles['diff-after']}>
-                <Form>
-                  {approvalInfo?.enableApproval && (
-                    <div className={styles.approval}>
-                      <div className={styles['approval-title']}>
-                        <FormattedMessage id="toggles.settings.approval.reviewers" />:
-                        <Popup
-                          inverted
-                          trigger={<Icon type="info" customclass={styles['icon-info']} />}
-                          content={intl.formatMessage({ id: 'targeting.approval.tips' })}
-                          position="top center"
-                          className="popup-override"
-                        />
-                      </div>
-                      <div className={styles['approval-content']}>
-                        <Dropdown
-                          fluid
-                          multiple
-                          value={approvalInfo?.reviewers}
-                          options={options}
-                          renderLabel={renderLabel}
-                          icon={null}
-                          disabled={true}
-                          className={styles['approval-dropdown']}
-                        />
-                        <div className={styles['approval-btn']} onClick={handleGotoSetting}>
-                          <FormattedMessage id="common.toggle.approval.settings.text" />
-                        </div>
-                      </div>
-                    </div>
-                  )}
-                  
-                  {
-                    !approvalInfo?.enableApproval && allowEnableTrackEvents && (
-                      <div className={styles.collect}>
-                        <div className={styles['collect-title']}>
-                          <span className="label-required">*</span>
-                          <FormattedMessage id='targeting.publish.start.collect' />
-                        </div>
-                        <div className={styles['collect-content']}>
-                          <Form.Field className={styles['collect-radio']}>
-                            <Form.Radio
-                              label={intl.formatMessage({id: 'common.yes.text'})}
-                              name='radioGroup'
-                              value='yes'
-                              checked={isCollect === 'yes'}
-                              error={newFormState.errors.radioGroup ? true : false}
-                              onChange={async (e: SyntheticEvent, detail: RadioProps) => {
-                                handleRadioChange(e, detail);
-                                newSetValue(detail.name || 'radioGroup', detail.value);
-                                await newTrigger('radioGroup');
-                              }}
-                            />
-                          </Form.Field>
-                          <Form.Field>
-                            <Form.Radio
-                              label={intl.formatMessage({id: 'common.no.text'})}
-                              name='radioGroup'
-                              value='no'
-                              checked={isCollect === 'no'}
-                              error={newFormState.errors.radioGroup ? true : false}
-                              onChange={async (e: SyntheticEvent, detail: RadioProps) => {
-                                handleRadioChange(e, detail);
-                                newSetValue(detail.name || 'radioGroup', detail.value);
-                                await newTrigger('radioGroup');
-                              }}
-                            />
-                          </Form.Field>
-                        </div>
-                        {newFormState.errors.radioGroup && (
-                          <div className="error-text">
-                            <FormattedMessage id="analysis.select.placeholder" />
-                          </div>
-                        )}
-                      </div>
-                    )
-                  }
-
-                  <div className={styles.comment}>
-                    <div className={styles['comment-title']}>
-                      {approvalInfo?.enableApproval && <span className="label-required">*</span>}
-                      <FormattedMessage id="targeting.publish.modal.comment" />:
-                    </div>
-                    <div className={styles['comment-content']}>
-                      <Form.TextArea
-                        name="reason"
-                        error={newFormState.errors.reason ? true : false}
-                        className={styles['comment-input']}
-                        placeholder={intl.formatMessage({ id: 'common.input.placeholder' })}
-                        onChange={async (e: SyntheticEvent, detail: TextAreaProps) => {
-                          handleInputComment(e, detail);
-                          newSetValue(detail.name, detail.value);
-                          await newTrigger('reason');
-                        }}
-                      />
-                      {newFormState.errors.reason && (
-                        <div className="error-text">
-                          <FormattedMessage id="common.input.placeholder" />
-                        </div>
-                      )}
-                    </div>
-                  </div>
-                </Form>
-              </div>
-            </div>
-          </div>
-        </Modal>
-        <Prompt when={!publishDisabled} message={intl.formatMessage({ id: 'targeting.page.leave.text' })} />
-        <Joyride
-          run={run}
-          callback={handleJoyrideCallback}
-          stepIndex={stepIndex}
-          continuous
-          hideCloseButton
-          scrollToFirstStep
-          showProgress={false}
-          showSkipButton
-          scrollOffset={100}
-          disableCloseOnEsc={true}
-          steps={STEPS}
-          locale={{
-            back: intl.formatMessage({ id: 'guide.last' }),
-            next: intl.formatMessage({ id: 'guide.next' }),
-            last: intl.formatMessage({ id: 'guide.done' }),
-          }}
-          floaterProps={{ ...floaterStyle }}
-          styles={{ ...tourStyle }}
+        <PublishModal
+          open={open}
+          approvalInfo={approvalInfo}
+          trackEvents={trackEvents}
+          allowEnableTrackEvents={allowEnableTrackEvents}
+          publishTargeting={publishTargeting}
+          initialTargeting={initialTargeting}
+          latestVersion={latestVersion}
+          setOpen={setOpen}
+          setLoading={setLoading}
+          initTargeting={initTargeting}
         />
+        <Prompt when={!publishDisabled} message={intl.formatMessage({ id: 'targeting.page.leave.text' })} />
+        <UserGuide />
       </Form>
       <SizeTips hide={sizeConfirm || sizeState === 'normal'} size={size} onConfirm={() => {
         setSizeConfirm(true);
